Makes python wrapper optionnal (old behavior conserved).
If an user does have python installed but numpy not installed, an
error will be shown.

The compiltaion of the wrapper should be optional (and maybe if set to
On should produce error). This patch just skip the build of the
wrapper if the user explicitely disable it. It defaults to the old behavior.