Improves build system for edge cases. (#111)
* Uses PROJECT_SOURCE_DIR instead of CMAKE_SOURCE_DIR

When trying to build apriltag using cmake `ExternalProject_Add` or
`FetchContent`, the build will fail as the python and tag macro does
not work anymore.

In such a case `CMAKE_SOURCE_DIR` refers to the parent cmake project
that uses `ExternalProject_Add` and therefore the python wrapper nor
the tag family source are found.

Using `PROJECT_SOURCE_DIR` solves the issue as it refers in any
situation to the apriltag project source dir, which I believ was the
intended behavior.

* Makes python wrapper optionnal (old behavior conserved).

If an user does have python installed but numpy not installed, an
error will be shown.

The compiltaion of the wrapper should be optional (and maybe if set to
On should produce error). This patch just skip the build of the
wrapper if the user explicitely disable it. It defaults to the old behavior.