Lift & shift demo/example code, and tracing to callbacks
This change takes the image debugging code and moves it into the example folder.

The image generation code is identical with the same inputs and outputs but instead
of existing inline with the program code it exists through a selectable macro.

This change facilities moving a lot of code not used by the core algorithm into a
support library so it's not neccessary to link this code with the core library leading
to a large code savings.

This also unlocks the ability to hook the various stages of the algorithm with real-time
visualizations by utilizing the same inputs as the existing debug code.