Speed up cluster merge by batch copying arrays
merge_clusters was using zarray_add_all which was copying over elements
one-by-one doing 2 memcpys and a potential array resize per element.
Here we replace it by a range copy that does a single resize and memcpy
for the operation which is a lot faster.

In my testing it reduces the total runtime for an image that's 2000x3000
by 20%.