apriltag_quad_thresh: Prevent using decimate for scale smaller than 1
The rest of the codebase treats all values equal to 1 or smaller as disablers of the decimation. This function was the only one behaving differently. After this commit, the behavior will be consistent.

Especially, value 0.0 led to division by zero and then conversion of NaN, to int, which is undefined behavior.

Signed-off-by: Martin Pecka<peckama2@fel.cvut.cz>