workerpool: Add predicate to condition variable startcond
Threads that wait on the condition variable startcond can be woken up
spuriously. This is not a problem if no tasks have been added, but as
soon as there is one, the condition in the while loop of worker_thread
will evaluate to false and the worker thread will run the task. If more
tasks are added, there is a chance that the zarray is reallocated and
that its pointer is changed. Since the woker thread gets a raw pointer
to the zarray, this pointer can become invalid and cause a segfault.

The solution is to add a predicate that is checked when the worker
thread is woken up on the condition variable startcond. The while loop
now has 2 conditions:
- start condition: if wp->start_tasks is true, tasks can be executed
- stop condition: if wp->taskspos == zarray_size(wp->tasks), there are
no more tasks to execute

I also shielded concurrent accesses to the pointer wp with the mutex.