Fix the inconsistent image coordinate conventions.
The convention that (0,0) be the left top corner of
the first pixel is adopted in this commit.

In the previous versions, the coordinate convention
used for `quad_decimate = 1, refine_edges = false` is
already `left-top-corner = (0,0)`;
while for `quad_decimate > 1` or `refine_edges = true`,
the convention is vague since it seems mixed conventions
are used and some code even be wrong (see
https://github.com/AprilRobotics/apriltag/issues/345).

This fix ensures the same convention be used no matter
what values for `quad_decimate` and `refine_edges`.